backend/requirements.txt
quart
quart_cors
openai
python-dotenv
pypdf2
langchain
tiktoken